requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==4.6.0
passlib==1.7.4
bcrypt==4.0.1
jinja2==3.1.2
pydantic==2.4.2
email-validator==2.1.0.post1
requests==2.32.3
